rename livisi_entity.py to entity.py
