+++ homeassistant/components/livisi/entity.py
