+++ homeassistant/components/livisi/switch.py
@@ -17,7 +17,7 @@ from .const import (
     PSS_DEVICE_TYPE,
 )
 from .coordinator import LivisiDataUpdateCoordinator
-from .livisi_entity import LivisiEntity
+from .entity import LivisiEntity
 
 
 async def async_setup_entry(
